feat: 服装に関する型定義の見直し
