+++ src/types/index.ts
@@ -3,6 +3,8 @@
  */
 
 export * from "./clothing";
+export * from "./clothingImg";
 export * from "./color";
 export * from "./forecast";
+export * from "./tempClothingRule";
 export * from "./time";
